fix: don't ignore schemas of contained resources
If a field appears in {"contained": [{my resource fields}]}, it will
now also appear in the resulting schema.

If multiple resources are in the contained list, we'll use all the
relevant resource schema to find the column schema for mentioned
fields. This is slightly janky, but it's not clear how else to
handle this without pulling resources out of the contained list.
A project for the future.