json: when reading the first line, use a much smaller block size
This helps speed up big file tree scans, especially over the network.
For example, the previous default for S3, was 50MB - now we just read
9kB.

In a sample scan of a full BCH archive tree on S3, a scan of all files
went from 21 minutes to 7. Still room for improvement there, but it's
better.